+++ tests/Api.UnitTests/Controllers/BeersControllerTests.cs
@@ -6,6 +6,9 @@ using Application.Beers.Dtos;
 using Application.Beers.Queries.GetBeer;
 using Application.Beers.Queries.GetBeers;
 using Application.Common.Models;
+using Application.Favorites.Commands.CreateFavorite;
+using Application.Favorites.Commands.DeleteFavorite;
+using Application.Favorites.Queries.GetFavorites;
 using Microsoft.AspNetCore.Mvc;
 using Moq;
 
@@ -120,7 +123,7 @@ public class BeersControllerTests : ControllerSetup<BeersController>
         var response = await Controller.UpdateBeer(id, command);
 
         // Assert
-        response.Should().BeOfType<BadRequestResult>();
+        response.Should().BeOfType<BadRequestObjectResult>().Which.Value.Should().Be(ExpectedInvalidIdMessage);
     }
 
     /// <summary>
@@ -136,6 +139,79 @@ public class BeersControllerTests : ControllerSetup<BeersController>
 
         // Act
         var response = await Controller.DeleteBeer(id);
+
+        // Assert
+        response.Should().BeOfType<NoContentResult>();
+    }
+
+    /// <summary>
+    ///     Tests that GetFavorites method returns PaginatedList of BeerDto.
+    /// </summary>
+    [Fact]
+    public async Task GetFavorites_ShouldReturnPaginatedListOfBeerDto()
+    {
+        // Arrange
+        var favoriteBeers = new List<BeerDto>
+        {
+            new()
+            {
+                Id = Guid.NewGuid(),
+                Name = "Test beer 1",
+                AlcoholByVolume = 6
+            },
+            new()
+            {
+                Id = Guid.NewGuid(),
+                Name = "Test beer 2",
+                AlcoholByVolume = 5
+            }
+        };
+        var expectedResult = PaginatedList<BeerDto>.Create(favoriteBeers, 1, 10);
+        var query = new GetFavoritesQuery();
+
+        MediatorMock.Setup(m => m.Send(query, CancellationToken.None)).ReturnsAsync(expectedResult);
+
+        // Act
+        var response = await Controller.GetFavorites(query);
+
+        // Assert
+        response.Result.Should().BeOfType<OkObjectResult>()
+            .Which.Value.Should().BeSameAs(expectedResult);
+        Controller.Response.Headers.Should().ContainKey("X-Pagination");
+        Controller.Response.Headers["X-Pagination"].Should().BeEquivalentTo(expectedResult.GetMetadata());
+    }
+
+    /// <summary>
+    ///     Tests that CreateFavorite method returns NoContent.
+    /// </summary>
+    [Fact]
+    public async Task CreateFavorite_ShouldReturnNoContent()
+    {
+        // Arrange
+        var beerId = Guid.NewGuid();
+        MediatorMock.Setup(m => m.Send(It.IsAny<CreateFavoriteCommand>(), CancellationToken.None))
+            .Returns(Task.CompletedTask);
+
+        // Act
+        var response = await Controller.CreateFavorite(beerId);
+
+        // Assert
+        response.Should().BeOfType<NoContentResult>();
+    }
+
+    /// <summary>
+    ///     Tests that DeleteFavorite method returns NoContent.
+    /// </summary>
+    [Fact]
+    public async Task DeleteFavorite_ShouldReturnNoContent()
+    {
+        // Arrange
+        var beerId = Guid.NewGuid();
+        MediatorMock.Setup(m => m.Send(It.IsAny<DeleteFavoriteCommand>(), CancellationToken.None))
+            .Returns(Task.CompletedTask);
+
+        // Act
+        var response = await Controller.DeleteFavorite(beerId);
 
         // Assert
         response.Should().BeOfType<NoContentResult>();
