--- tests/Api.UnitTests/Controllers/FavoritesControllerTests.cs
@@ -1,80 +1,0 @@
-﻿using Api.Controllers;
-using Application.Beers.Dtos;
-using Application.Common.Mappings;
-using Application.Favorites.Commands.CreateFavorite;
-using Application.Favorites.Commands.DeleteFavorite;
-using Application.Favorites.Dtos;
-using Application.Favorites.Queries.GetFavorites;
-using Microsoft.AspNetCore.Mvc;
-using Moq;
-
-namespace Api.UnitTests.Controllers;
-
-/// <summary>
-///     Unit tests for the <see cref="FavoritesController"/> class.
-/// </summary>
-[ExcludeFromCodeCoverage]
-public class FavoritesControllerTests : ControllerSetup<FavoritesController>
-{
-    /// <summary>
-    ///     Tests that GetFavorites method returns FavoritesListDto.
-    /// </summary>
-    [Fact]
-    public async Task GetFavorites_ShouldReturnFavoritesListDto()
-    {
-        // Arrange
-        var favoriteBeers = new List<BeerDto>().ToPaginatedList(1, 10);
-        var expectedResult = new FavoritesListDto
-        {
-            UserId = Guid.NewGuid(),
-            FavoriteBeers = favoriteBeers
-        };
-        var query = new GetFavoritesQuery();
-
-        MediatorMock.Setup(m => m.Send(query, CancellationToken.None)).ReturnsAsync(expectedResult);
-
-        // Act
-        var response = await Controller.GetFavorites(query);
-
-        // Assert
-        response.Result.Should().BeOfType<OkObjectResult>()
-            .Which.Value.Should().BeSameAs(expectedResult);
-        Controller.Response.Headers.Should().ContainKey("X-Pagination");
-        Controller.Response.Headers["X-Pagination"].Should().BeEquivalentTo(expectedResult.FavoriteBeers.GetMetadata());
-    }
-
-    /// <summary>
-    ///     Tests that CreateFavorite method returns NoContent.
-    /// </summary>
-    [Fact]
-    public async Task CreateFavorite_ShouldReturnNoContent()
-    {
-        // Arrange
-        var command = new CreateFavoriteCommand { BeerId = Guid.NewGuid() };
-        MediatorMock.Setup(m => m.Send(command, CancellationToken.None)).Returns(Task.CompletedTask);
-
-        // Act
-        var response = await Controller.CreateFavorite(command);
-
-        // Assert
-        response.Should().BeOfType<NoContentResult>();
-    }
-
-    /// <summary>
-    ///     Tests that DeleteFavorite method returns NoContent.
-    /// </summary>
-    [Fact]
-    public async Task DeleteFavorite_ShouldReturnNoContent()
-    {
-        // Arrange
-        var id = Guid.NewGuid();
-        MediatorMock.Setup(m => m.Send(It.IsAny<DeleteFavoriteCommand>(), CancellationToken.None))
-            .Returns(Task.CompletedTask);
-
-        // Act
-        var response = await Controller.DeleteFavorite(id);
-
-        // Assert
-        response.Should().BeOfType<NoContentResult>();
-    }
-}
